More than year of works from a master branch are here

+++ Engine/common_features/fader.cpp
@@ -1,24 +1,26 @@
 /*
- * Platformer Game Engine by Wohlstand, a free platform for game making
- * Copyright (c) 2017 Vitaly Novichkov <admin@wohlnet.ru>
+ * Moondust, a free game engine for platform game making
+ * Copyright (c) 2014-2020 Vitaly Novichkov <admin@wohlnet.ru>
  *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * any later version.
+ * This software is licensed under a dual license system (MIT or GPL version 3 or later).
+ * This means you are free to choose with which of both licenses (MIT or GPL version 3 or later)
+ * you want to use this software.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * You can see text of MIT license in the LICENSE.mit file you can see in Engine folder,
+ * or see https://mit-license.org/.
+ *
+ * You can see text of GPLv3 license in the LICENSE.gpl3 file you can see in Engine folder,
+ * or see <http://www.gnu.org/licenses/>.
  */
 
 #include "fader.h"
 #include <stdlib.h>
 #include <math.h>
+#include <Utils/maths.h>
 
 PGE_Fader::PGE_Fader()
 {
@@ -58,16 +60,16 @@ void PGE_Fader::setRatio(double ratio)
 bool PGE_Fader::tickFader(double ticks)
 {
     if(fadeSpeed < 1) return true; //Idling animation
-    if(cur_ratio==target_ratio)
+    if(cur_ratio == target_ratio)
         return true;
 
     manual_ticks -= fabs(ticks);
-    while(manual_ticks<=0.0)
+    while(manual_ticks <= 0.0)
     {
         fadeStep();
         manual_ticks += fadeSpeed;
     }
-    return (cur_ratio==target_ratio);
+    return (cur_ratio == target_ratio);
 }
 
 const double &PGE_Fader::fadeRatio()
@@ -77,12 +79,12 @@ const double &PGE_Fader::fadeRatio()
 
 bool PGE_Fader::isFading()
 {
-    return (cur_ratio!=target_ratio);
+    return (cur_ratio != target_ratio);
 }
 
 bool PGE_Fader::isFull()
 {
-    return ((cur_ratio>=1.0) && (!isFading()));
+    return ((cur_ratio >= 1.0) && (!isFading()));
 }
 
 void PGE_Fader::setFull()
@@ -93,7 +95,7 @@ void PGE_Fader::setFull()
 
 bool PGE_Fader::isNull()
 {
-    return ((cur_ratio<=0.0) && (!isFading()));
+    return ((cur_ratio <= 0.0) && (!isFading()));
 }
 
 void PGE_Fader::setNull()
@@ -104,19 +106,18 @@ void PGE_Fader::setNull()
 
 int PGE_Fader::ticksLeft()
 {
-    return manual_ticks;
+    return static_cast<int>(manual_ticks);
 }
 
 void PGE_Fader::fadeStep()
 {
-    if(cur_ratio==target_ratio)
+    if(cur_ratio == target_ratio)
         return;
     if(cur_ratio < target_ratio)
-        cur_ratio+=fade_step;
+        cur_ratio += fade_step;
     else
-        cur_ratio-=fade_step;
+        cur_ratio -= fade_step;
 
-    if(cur_ratio>1.0) cur_ratio = 1.0;
-    else
-        if(cur_ratio<0.0) cur_ratio = 0.0;
+    if(cur_ratio > 1.0) cur_ratio = 1.0;
+    else if(cur_ratio < 0.0) cur_ratio = 0.0;
 }
